Update rocket.h (#5)
I added comments to the rocket.h file, although, I don't completely understand how some of the functions work so if any of y'all know what it's supposed to do just tell me what it does and I'll fix the comments. I'll have to the add the comments in the cpp file since we need comments on every file.