Add files via upload (#15)
bullets.h and bullets.cpp
The rocket.h and rocket.cpp file also has some stuff regarding bullets (needs to be updated). The bullets do shoot from the rocket but the bullets need to be larger and the collision detection is not set up yet. The files will be updated soon.